fix: re-organise folder structure to remove test data from made seed file
