+++ server/prisma/db-data/index.ts
@@ -1,0 +1,2 @@
+export * from './organisations';
+export * from './departments';
