+++ server/prisma/db-data/organisations.ts
@@ -1,0 +1,24 @@
+const organisationData = [
+  {
+    id: '67877aab-e235-431a-b960-8da259414d0c',
+    name: 'Coca-Cola Inc.'
+  }, 
+  {
+    id: '3bad9076-dfb0-439d-8e36-9406909bedbb',
+    name: 'Pepsi Co'
+  },
+  {
+    id: '2e37ffc4-008c-4640-9130-0980de1fb648',
+    name: 'Walmart'
+  },
+  {
+    id: '0e3bf95c-6107-4837-9878-f05d1b7286ed',
+    name: 'Mcdonalds'
+  },
+  {
+    id: 'da5985e5-c0c2-44a7-bfd4-e3b5ebf452c0',
+    name: 'Amazon'
+  }
+]
+
+export { organisationData };
